pyproject.toml
[tool.poetry]
package-mode = false
authors = ["Vladislav Romaschenko <vlad1vaselkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["uvicorn"], version = "^0.111.0"}
pydantic = {extras = ["email"], version = "^2.7.3"}
pika = "^1.3.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.2.2"
aiosmtpd = "^1.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
